Duel model/controller base, cpu actions, How Can You See??
Notes:

This is a massive update and I didn't write or perform any runtime tests yet.

DuelController need to have duelID correctly set to work. It should be
set to one of the values in the enum DuelID, which correspond to the
battle stage.

There are three main public methods.

initDuel()
playerChoosesSpecial( Special s )
playerChoosesCard( Card c )

Right now we assume initDuel will be called by the DuelUIController
before we do anything else. The last two correspond to button triggers.

Certain parts may be redundant with with is done in DuelUIController,
and these are noted with TODO comments so we can review and remove them if
necessary.

How Can You See?? is implemented by setting the public boolean property
enemyChoiceVisible to true. DuelUIController can use this to know to
update the view.

useSpecial( Special s ) is a method with a large switch statement that
recognizes the names of all the Specials. The Special name has to be set
correctly.

runSpecials() runs all the specials at a particular timing point. Timing
for the Special needs to be set correctly.

0 = immediately/before player action
1 = after damage calculation